requirements.txt
cohere==5.13.4
faiss_cpu==1.8.0
Flask==3.1.0
groq==0.18.0
gunicorn==21.2.0
huggingface_hub==0.26.3
nltk==3.8.1
numpy==1.23.5
python-dotenv==1.0.1
rank_bm25==0.2.2
scikit_learn==1.1.3
sentence_transformers==3.0.1
torch==2.5.1
transformers==4.46.3
Werkzeug==3.1.3